move fraud proof verification to sp-domains-fraud-proof and storage proof provider and verifier to its own module
