+++ crates/sp-domains/src/lib.rs
@@ -21,12 +21,12 @@ pub mod bundle_producer_election;
 pub mod extrinsics;
 pub mod fraud_proof;
 pub mod merkle_tree;
+pub mod proof_provider_and_verifier;
 pub mod storage;
 #[cfg(test)]
 mod tests;
 pub mod transaction;
-pub mod valued_trie_root;
-pub mod verification;
+pub mod valued_trie;
 
 extern crate alloc;
